+++ SkillsLab2023_Assignment/Models/EditTrainingViewModel.cs
@@ -9,7 +9,7 @@ namespace SkillsLab2023_Assignment.Models
     public class EditTrainingViewModel : TrainingViewModel
     {
         [Required(ErrorMessage = "TrainingId is required")]
-        public int TrainingId { get; set; }
+        public short TrainingId { get; set; }
 
     }
 }
